Implement Context reducer & Start Context cleanups
