+++ src/components/PadEditorItem/EditorItemButtons.js
@@ -1,4 +1,6 @@
-import React from 'react';
+import React, {useContext} from 'react';
+import {updateEditorData} from '../../actions'
+import {Context} from '../../contexts/SamplerContext';
 import Colors from '../../Config/ColorScheme';
 import './PadEditorButtons.css';
 
@@ -15,19 +17,20 @@ let style = {
     margin: "1vh auto"
 }
 
-export default ({cmd1, cmd2, content1, content2, val1, val2, updateEditorData}) => {
+export default ({cmd1, cmd2, content1, content2, val1, val2}) => {
+    const context = useContext(Context)
     return (
         <div className="pad-item-wrapper" style={style}>
             <button
             className="editor-btn"
             id="prev"
-            onClick={(e) => { updateEditorData({cmd: cmd1, val: val1})}}>
+            onClick={(e) => { updateEditorData({context, cmd: cmd1, val: val1})}}>
             {content1}
             </button>
             <button
             className="editor-btn"
             id="next"
-            onClick={(e) => { updateEditorData({cmd: cmd2, val: val2})}}>
+            onClick={(e) => { updateEditorData({context, cmd: cmd2, val: val2})}}>
             {content2}
             </button>
         </div>
