+++ src/components/PadEditorItem/EditorItemRange.js
@@ -1,4 +1,6 @@
-import React from 'react';
+import React, {useContext} from 'react';
+import {updateEditorData} from '../../actions'
+import {Context} from '../../contexts/SamplerContext';
 import Colors from '../../Config/ColorScheme';
 import './PadEditorItem.css';
 
@@ -16,6 +18,7 @@ let style = {
 }
 
 export default (props) => {
+    const context = useContext(Context)
     return (
         <div className="pad-item-wrapper" style={style}>
             <label htmlFor={props.label} 
@@ -29,7 +32,7 @@ export default (props) => {
             name={props.label} 
             step={props.step}
             id={props.label}
-            onChange={(e) => { props.updateEditorData({cmd: props.label, val: e.target.value})}}/>
+            onChange={(e) => { updateEditorData({context, cmd: props.label, val: e.target.value})}}/>
         </div>
     )
 }
